refactor: split harness daemon modules
- Keep index.ts as the thin CLI entrypoint
- Move inventory, shell, tmux, worktree, spawner, command, and type logic into focused modules
- Update the plan doc with the module split
