+++ extensions/harness-daemon/src/errors.ts
@@ -1,0 +1,3 @@
+export function die(message: string): never {
+  throw new Error(message)
+}
